+++ src/Request/Internal.php
@@ -367,7 +367,8 @@ class Internal extends RequestCollection
                 $request
                     ->addPost('client_shared_at', (string) time())
                     ->addPost('source_type', '3')
-                    ->addPost('configure_mode', 1)
+                    ->addPost('configure_mode', '1')
+                    ->addPost('allow_multi_configures', '1')
                     //->addPost('configure_mode', Constants::SHARE_TYPE['REEL_SHARE']) // 2 - REEL_SHARE
                     ->addPost('client_timestamp', (string) (time() - mt_rand(3, 10)))
                     ->addPost('upload_id', $uploadId)
@@ -375,7 +376,9 @@ class Internal extends RequestCollection
                     ->addPost('scene_capture_type', '')
                     ->addPost('creation_surface', 'camera')
                     ->addPost('capture_type', 'normal')
-                    ->addPost('has_original_sound', '1');
+                    ->addPost('has_original_sound', '1')
+                    ->addPost('composition_id', Signatures::generateUUID())
+                    ->addPost('camera_entry_point', '11');
 
                 if (is_string($link) && Utils::hasValidWebURLSyntax($link)) {
                     $story_cta = '[{"links":[{"webUri":'.json_encode($link).'}]}]';
@@ -963,11 +966,14 @@ class Internal extends RequestCollection
                 }
 
                 $request
-                    ->addPost('configure_mode', 1)
+                    ->addPost('configure_mode', '1')
                     //->addPost('configure_mode', Constants::SHARE_TYPE['REEL_SHARE']) // 2 - REEL_SHARE
+                    ->addPost('allow_multi_configures', '1')
                     ->addPost('story_media_creation_date', time() - mt_rand(10, 20))
                     ->addPost('client_shared_at', time() - mt_rand(3, 10))
-                    ->addPost('client_timestamp', time());
+                    ->addPost('client_timestamp', time())
+                    ->addPost('composition_id', Signatures::generateUUID())
+                    ->addPost('camera_entry_point', '11');
 
                 if (is_string($link) && Utils::hasValidWebURLSyntax($link)) {
                     $story_cta = '[{"links":[{"webUri":'.json_encode($link).'}]}]';